Add bots Pine and Zakiyama, reintroduce SetBotScore method
Refactored code to remove the addition of the bot "Artorius" and
add bots "Pine" and "Zakiyama" with specific difficulty settings.
Reintroduced the `SetBotScore` method, moving it to a different
location in the code. The method is now called immediately after
setting up the bot's weapon services and sets scores for "Pine"
and "Zakiyama" to 999 and 888 respectively. Added a comment to
the `GetDefaultPrimaryWeapon` method indicating that special
bots do not have a weapon.