Add SetClientModel method and refactor code to use it
The `Utility.cs` file has been updated to include a new `SetClientModel` method. This method sets the model for a client using a specified skin name, retrieves the skin from a dictionary, and applies the model path to the client's player pawn. If the skin has a mesh group index, it also sets the mesh group mask using the `ComputeMeshGroupMask` method.

The `Bot.cs` and `Command.cs` files have been refactored to use the new `SetClientModel` method instead of directly setting the model and mesh group mask for player pawns. This change simplifies the code and reduces redundancy.

An additional using directive for `CounterStrikeSharp.API.Core` has been added to `Utility.cs`.